Move the buttons into a static toolbar
