+++ chemclipse/plugins/org.eclipse.chemclipse.ux.extension.xxd.ui/fragment.e4xmi
@@ -2,12 +2,7 @@
 <fragment:ModelFragments xmi:version="2.0" xmlns:xmi="http://www.omg.org/XMI" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:advanced="http://www.eclipse.org/ui/2010/UIModel/application/ui/advanced" xmlns:basic="http://www.eclipse.org/ui/2010/UIModel/application/ui/basic" xmlns:basic_1="http://www.eclipse.org/ui/2010/UIModel/application/descriptor/basic" xmlns:commands="http://www.eclipse.org/ui/2010/UIModel/application/commands" xmlns:fragment="http://www.eclipse.org/ui/2010/UIModel/fragment" xmlns:menu="http://www.eclipse.org/ui/2010/UIModel/application/ui/menu" xmi:id="_RvTHsKuDEeSKQ7OYnQVHjQ">
   <imports xsi:type="advanced:Area" xmi:id="_qjvtkCskEeakPPc3RJFSAg" elementId="org.eclipse.chemclipse.rcp.app.ui.editor"/>
   <fragments xsi:type="fragment:StringModelFragment" xmi:id="_TxCZIKuHEeSKQ7OYnQVHjQ" featurename="sharedElements" parentElementId="org.eclipse.chemclipse.rcp.app.ui.trimmedwindow.main">
-    <elements xsi:type="basic:Part" xmi:id="_qtur4Cf2EeafjOXnfHSw0Q" elementId="org.eclipse.chemclipse.ux.extension.xxd.ui.part.dataexplorer" contributionURI="bundleclass://org.eclipse.chemclipse.ux.extension.xxd.ui/org.eclipse.chemclipse.ux.extension.xxd.ui.parts.DataExplorerPart" label="Data Explorer" iconURI="platform:/plugin/org.eclipse.chemclipse.rcp.ui.icons/icons/16x16/chromatogramFileExplorer.gif">
-      <toolbar xmi:id="_0M9cAIavEemwu6_7mTaq4w" elementId="org.eclipse.chemclipse.ux.extension.xxd.ui.toolbar.0">
-        <children xsi:type="menu:DirectToolItem" xmi:id="_QFr2YIawEemwu6_7mTaq4w" elementId="org.eclipse.chemclipse.ux.extension.xxd.ui.directtoolitem.reset" label="Reset" iconURI="platform:/plugin/org.eclipse.chemclipse.rcp.ui.icons/icons/16x16/reset.gif" tooltip="Reset the file explorer" contributionURI="bundleclass://org.eclipse.chemclipse.ux.extension.xxd.ui/org.eclipse.chemclipse.ux.extension.xxd.ui.internal.handlers.ResetDataExplorerHandler"/>
-        <children xsi:type="menu:DirectToolItem" xmi:id="_Tru6kIaxEemwu6_7mTaq4w" elementId="org.eclipse.chemclipse.ux.extension.xxd.ui.directtoolitem.settings" label="Settings" iconURI="platform:/plugin/org.eclipse.chemclipse.rcp.ui.icons/icons/16x16/configure.gif" tooltip="Open the Settings" contributionURI="bundleclass://org.eclipse.chemclipse.ux.extension.xxd.ui/org.eclipse.chemclipse.ux.extension.xxd.ui.internal.handlers.DataExplorerSettingsHandler"/>
-      </toolbar>
-    </elements>
+    <elements xsi:type="basic:Part" xmi:id="_qtur4Cf2EeafjOXnfHSw0Q" elementId="org.eclipse.chemclipse.ux.extension.xxd.ui.part.dataexplorer" contributionURI="bundleclass://org.eclipse.chemclipse.ux.extension.xxd.ui/org.eclipse.chemclipse.ux.extension.xxd.ui.parts.DataExplorerPart" label="Data Explorer" iconURI="platform:/plugin/org.eclipse.chemclipse.rcp.ui.icons/icons/16x16/chromatogramFileExplorer.gif"/>
     <elements xsi:type="basic:Part" xmi:id="_dXp1AC5uEeiL24_JU7_FGA" elementId="org.eclipse.chemclipse.ux.extension.xxd.ui.part.peakScanListPart" contributionURI="bundleclass://org.eclipse.chemclipse.ux.extension.xxd.ui/org.eclipse.chemclipse.ux.extension.xxd.ui.parts.PeakScanListPart" label="Peak/Scan List" iconURI="platform:/plugin/org.eclipse.chemclipse.rcp.ui.icons/icons/16x16/peaks.gif" tooltip=""/>
     <elements xsi:type="basic:Part" xmi:id="_E8UhsDmqEei-wI9hy1kIxA" elementId="org.eclipse.chemclipse.ux.extension.xxd.ui.part.targets" contributionURI="bundleclass://org.eclipse.chemclipse.ux.extension.xxd.ui/org.eclipse.chemclipse.ux.extension.xxd.ui.parts.TargetsPart" label="Targets" iconURI="platform:/plugin/org.eclipse.chemclipse.rcp.ui.icons/icons/16x16/target.png"/>
     <elements xsi:type="basic:Part" xmi:id="_NC1V0DmqEei-wI9hy1kIxA" elementId="org.eclipse.chemclipse.ux.extension.xxd.ui.part.comparisonScan" contributionURI="bundleclass://org.eclipse.chemclipse.ux.extension.xxd.ui/org.eclipse.chemclipse.ux.extension.xxd.ui.parts.ComparisonScanChartPart" label="Comparison Scan" iconURI="platform:/plugin/org.eclipse.chemclipse.rcp.ui.icons/icons/16x16/comparisonScanDefault.gif"/>
